Fixed TypeCast-Error and support network check on app comes to front
The property self.webView should be casted to UIWebView otherwise you will get a type cast error.  Also there is a problem, that when the offline view comes visible and the app runs in background it will not recieve connection changed notification. Thiswhy I added a handler to react on the event, when app is coming to foreground and force a network check at beginning.